refactor: extract email validation for sharing
