+++ app/routes/vet/email.js
@@ -4,7 +4,7 @@ const { notify: { templateIdVetLogin } } = require('../../config')
 const sendMagicLinkEmail = require('../../lib/email/send-magic-link-email')
 const session = require('../../session')
 const { vetSignup: { email: emailKey } } = require('../../session/keys')
-const { email: emailErrorMessages } = require('../../../app/lib/error-messages')
+const { email: emailValidation } = require('../../../app/lib/validation/email')
 
 module.exports = [{
   method: 'GET',
@@ -23,13 +23,7 @@ module.exports = [{
     auth: false,
     validate: {
       payload: Joi.object({
-        email: Joi.string().trim().email().required()
-          .messages({
-            'any.required': emailErrorMessages.enterEmail,
-            'string.base': emailErrorMessages.enterEmail,
-            'string.email': emailErrorMessages.validEmail,
-            'string.empty': emailErrorMessages.enterEmail
-          })
+        email: emailValidation
       }),
       failAction: async (request, h, error) => {
         return h.view('vet/email', { ...request.payload, errorMessage: { text: error.details[0].message } }).code(400).takeover()
